requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.25.0
coverage>=6.0
flake8>=5.0.0
flask-testing>=0.8.0